requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
canvas==0.3
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
flake8==6.1.0
idna==3.4
inflection==0.3.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.19
Pillow==10.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pytest==7.4.2
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rabbitmq==0.2.0
reportlab==4.0.4
requests==2.31.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
